+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/manual/service/ServiceTraffic.java
@@ -66,7 +66,7 @@ public class ServiceTraffic extends Metrics {
     @Setter
     @Getter
     @Column(name = NAME)
-    @ElasticSearch.Column(columnAlias = "name")
+    @ElasticSearch.Column(legacyName = "name")
     @ElasticSearch.MatchQuery
     @BanyanDB.SeriesID(index = 1)
     private String name = Const.EMPTY_STRING;